services/auth/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
asyncpg